+++ templates/api.h
@@ -149,7 +149,7 @@ $(proc)_platform_t *$(proc)_platform($(proc)_state_t *state);
 /* simulation functions */
 $(proc)_sim_t *$(proc)_new_sim($(proc)_state_t *state, $(proc)_address_t start_addr, $(proc)_address_t exit_addr);
 $(proc)_inst_t *$(proc)_next_inst($(proc)_sim_t *sim);
-unsigned int $(proc)_run_and_count_inst($(proc)_sim_t *sim);
+uint64_t $(proc)_run_and_count_inst($(proc)_sim_t *sim);
 void $(proc)_run_sim($(proc)_sim_t *sim);
 void $(proc)_step($(proc)_sim_t *sim);
 void $(proc)_delete_sim($(proc)_sim_t *sim);
